Move preview to fragment
