+++ app/src/main/res/layout/activity_widget_settings.xml
@@ -18,15 +18,6 @@
         app:titleTextAppearance="@style/ActionBarTitleText" />
 
     <FrameLayout
-        android:id="@+id/preview_container"
-        android:layout_width="match_parent"
-        android:layout_height="225dp"
-        android:padding="44dp">
-        <include
-            layout="@layout/widget_2x1"/>
-    </FrameLayout>
-
-    <FrameLayout
         android:id="@+id/fragment_container"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
