+++ app/src/main/res/layout/fragment_widget_settings.xml
@@ -3,12 +3,27 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/default_bg"
     >
 
   <LinearLayout
       android:layout_width="match_parent"
       android:layout_height="wrap_content"
+      android:orientation="vertical"
+      >
+
+    <FrameLayout
+        android:id="@+id/preview_container"
+        android:layout_width="match_parent"
+        android:layout_height="250dp"
+        android:padding="44dp">
+      <include
+          layout="@layout/widget_2x1"/>
+    </FrameLayout>
+
+  <LinearLayout
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:background="@color/default_bg"
       android:orientation="vertical"
       >
 
@@ -173,7 +188,6 @@
           />
 
     </LinearLayout>
-
   </LinearLayout>
-
+  </LinearLayout>
 </ScrollView>
