+++ app/src/main/res/values/styles.xml
@@ -121,7 +121,12 @@
     <item name="android:windowBackground">@drawable/bg_splash</item>
   </style>
 
-  <style name="ParanormalActivityTheme" parent="AppTheme"/>
+  <style name="ParanormalActivityTheme" parent="AppTheme">
+    <item name="android:navigationBarColor">@color/nav_bar_bg_2</item>
+    <item name="android:windowBackground">@android:color/transparent</item>
+    <item name="android:colorBackgroundCacheHint">@null</item>
+    <item name="android:windowShowWallpaper">true</item>
+  </style>
 
   <style name="NewsFeedActivityTheme" parent="AppTheme"/>
