+++ app/version.properties
@@ -1,2 +1,2 @@
-versionName=3.9.741
-versionCode=300900741
+versionName=3.9.742
+versionCode=300900742
